milestone7/requirements.txt
google-generativeai>=0.7.2
groq>=0.9.0
openai>=1.30.1
httpx==0.27.2
datasets>=2.19.1
pandas
numpy
scikit-learn
tqdm
plotly
streamlit>=1.35.0
pyngrok
requests
